mentalTech_Python/requirements.txt
fastapi==0.78.0
SQLAlchemy==1.4.39
uvicorn==0.18.2
psycopg2-binary  # ou asyncpg se você estiver usando asyncio
python-dotenv
pylint
pytest
httpx
pytest-asyncio
coverage
python-jose
passlib
python-multipart
requests
mock-alchemy
authlib
httpx
fastapi-sessions